Rename swich package to switchcase
